requirements.txt
streamlit==1.35.0
requests==2.32.3
python-dotenv==1.0.1
spacy==3.7.2
docx2txt==0.8
PyMuPDF==1.23.9

# Auto-install the SpaCy English model during deployment
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl

